STPK-1429 Support floating point sampling rates
The Kionix accelerometer found on the MRD 7 advertises
support for 6.25 and 12.5 Hz, so open up a little the
scope of what we consider a valid sampling rate.

Some %f specifications were replaced by %g because %f
shows zero digits in the style of 3.140000, and some
sysfs writes would not be accepted because of this format.

Issue: STPK-1429

+++ common.h
@@ -105,7 +105,7 @@ struct sensor_info_t
 	float scale;    /*default: 1. when set to 0, use channel specific value*/
 	float illumincalib;	/* to set the calibration for the ALS */
 
-	int sampling_rate;	/* requested events / second */
+	float sampling_rate;	/* requested events / second */
 
 	int dev_num;	/* Associated iio dev num, ex: 3 for /dev/iio:device3 */
 	int enable_count;
